Refactor session management to use SessionRegistry in NTCP2

+++ lib/transport/ntcp2/transport_sa2_invariant_test.go
@@ -22,7 +22,7 @@ func TestSA2_InvariantInitial(t *testing.T) {
 	count := transport.GetSessionCount()
 	mismatch := transport.ValidateSessionCountingInvariant()
 
-	require.Equal(t, 0, count, "Should start with zero sessions")
+	require.Equal(t, int32(0), count, "Should start with zero sessions")
 	require.Equal(t, 0, mismatch, "SA-2 invariant should hold at startup")
 }
